restructure project; move header into subdirectory; create CMakeLists in src sub directory
